Introduce tagged_with_tags_count
For example


	Photos.tagged_with(["funny", "naughty", "green"]).where{tags_count >= 2}

is the same as

	Photos.tagged_with(["funny", "naughty", "green"], :min => 2)

but more complex queries can be chained with the new first form. As
well the inner part of the query which contains group and
aggregate function is shielded from ActiveRelation so that
you can chain further aggregate functions if you wish in
an intuitive way.